+++ cogito/gchatsink.go
@@ -40,13 +40,12 @@ func (sink GoogleChatSink) Send() error {
 		return nil
 	}
 
+	// Prepare the message.
+	threadKey := fmt.Sprintf("%s %s", sink.Request.Env.BuildPipelineName, sink.GitRef)
+	text := prepareChatMessage(sink.Request, sink.GitRef)
+
 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
 	defer cancel()
-
-	pipeline := sink.Request.Env.BuildPipelineName
-	threadKey := fmt.Sprintf("%s %s", pipeline, sink.GitRef)
-	text := gChatFormatText(sink.GitRef, state, sink.Request.Source, sink.Request.Env)
-
 	reply, err := googlechat.TextMessage(ctx, webHook, threadKey, text)
 	if err != nil {
 		return fmt.Errorf("GoogleChatSink: %s", err)
@@ -68,8 +67,22 @@ func shouldSendToChat(state BuildState, notifyOnStates []BuildState) bool {
 	return false
 }
 
-// gChatFormatText returns a plain text message to be sent to Google Chat.
-func gChatFormatText(gitRef string, state BuildState, src Source, env Environment) string {
+// prepareChatMessage returns a message ready to be sent to the chat sink.
+func prepareChatMessage(request PutRequest, gitRef string) string {
+	var text string
+	if request.Params.ChatMessage != "" {
+		text = request.Params.ChatMessage
+	} else {
+		text = gChatBuildSummaryText(gitRef, request.Params.State, request.Source,
+			request.Env)
+	}
+
+	return text
+}
+
+// gChatBuildSummaryText returns a plain text message to be sent to Google Chat.
+func gChatBuildSummaryText(gitRef string, state BuildState, src Source, env Environment,
+) string {
 	now := time.Now().Format("2006-01-02 15:04:05 MST")
 
 	// Google Chat format for links with alternate name: