+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Google Chat: allow selecting which states trigger a notification to the chat, via configuration `source.notify_on_states`. Default (as before this tunable): [`abort`, `error`, `failure`].
 - Google Chat: enrich cogito logging with human-readable information about the chat space.
-- Google Chat: allow to specify a different chat space/webhook during put. 
+- Google Chat: allow to specify a different chat space/webhook during put (see `params.gchat_webhook`).
+- Google Chat: allow to specify a custom chat message, overriding the default build summary (see `params.chat_message`).
 
 ## [v0.8.0] - 2022-08-11